+++ WarhammerArmyAssembler/Option.cs
@@ -28,7 +28,7 @@ namespace WarhammerArmyAssembler
         public bool Multiple { get; set; }
         
 
-        public int Points { get; set; }
+        public double Points { get; set; }
         public bool PerModel { get; set; }
 
         public string PointsView { get; set; }
